agent/pyproject.toml
[project]
name = "revogue-agent"
version = "0.1.0"
description = "ReVogue LiveKit text chat agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents~=1.4",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
